feat: use custom hooks
