+++ src/components/Usuarios.tsx
@@ -1,16 +1,8 @@
-import React, { useEffect, useRef, useState } from 'react';
-import { reqResApi } from '../api/reqRes';
-import { ResResListado, Usuario } from '../interface/reqRes';
+import { Usuario } from '../interface/reqRes';
+import useUsers from '../hooks/useUsers';
 
 export const Usuarios = () => {
-
-    const [usuarios, setUsuarios] = useState<Usuario[]>([]);
-    useEffect(() => {
-        // llamado al API
-        CargarUsuarios();
-    }, [])
-
-    const paginaRef = useRef(1);
+    const {  usuarios, paginaAnterior,paginaSiguiente } = useUsers();
 
     const renderItem = (usuario: Usuario) => {
         return (
@@ -31,26 +23,6 @@ export const Usuarios = () => {
             </tr>
         )
     }
-
-    const CargarUsuarios = async () => {
-        const response = await reqResApi.get<ResResListado>('/users', {
-            params: {
-                page: paginaRef.current
-            }
-        })
-        // For the moment not use the error handling 
-        // .then(resp => {
-        //     setUsuarios(resp.data.data)
-        // })
-        // .catch(console.log);
-        if (response.data.data.length > 0) {
-            setUsuarios(response.data.data)
-            paginaRef.current++;
-        } else {
-            alert('No hay mas registros');
-        }
-    }
-
     return (
         <>
             <h3>Usuarios:</h3>
@@ -69,13 +41,14 @@ export const Usuarios = () => {
             </table>
             <button
                 className="btn  btn-primary"
+                onClick={paginaAnterior}
             >
                 Anterior
             </button>
             &nbsp;
             <button
                 className="btn  btn-primary"
-                onClick={CargarUsuarios}
+                onClick={paginaSiguiente}
             >
                 Siguiente
             </button>
